+++ src/wai/annotations/xdc/od_to_ic/__init__.py
@@ -1,0 +1,3 @@
+"""
+Package for the object-detection -> image-classification cross-domain converter.
+"""