+++ src/wai/annotations/xdc/od_to_ic/component/__init__.py
@@ -1,0 +1,1 @@
+from ._OD2ICXDC import OD2ICXDC