+++ src/wai/annotations/xdc/od_to_ic/specifier/__init__.py
@@ -1,0 +1,1 @@
+from ._OD2ICXDCSpecifier import OD2ICXDCSpecifier